split redux files
